+++ OUDS/Core/OUDS/Tests/OUDSTheme/MockTheme/MockTheme+SizingSemanticTokens.swift
@@ -22,30 +22,30 @@ extension MockTheme {
 
     // MARK: Semantic token - Sizing - Width height - Icon decorative
 
-    override var sizeWidthHeightIconShortest: SizingWidthHeightSemanticToken { Self.mockThemeSizeRawToken }
-    override var sizeWidthHeightIconShorter: SizingWidthHeightSemanticToken { Self.mockThemeSizeRawToken }
-    override var sizeWidthHeightIconShort: SizingWidthHeightSemanticToken { Self.mockThemeSizeRawToken }
-    override var sizeWidthHeightIconMedium: SizingWidthHeightSemanticToken { Self.mockThemeSizeRawToken }
-    override var sizeWidthHeightIconTall: SizingWidthHeightSemanticToken { Self.mockThemeSizeRawToken }
-    override var sizeWidthHeightIconTaller: SizingWidthHeightSemanticToken { Self.mockThemeSizeRawToken }
-    override var sizeWidthHeightIconTallest: SizingWidthHeightSemanticToken { Self.mockThemeSizeRawToken }
+    override var sizeIconDecorativeShortest: SizingWidthHeightSemanticToken { Self.mockThemeSizeRawToken }
+    override var sizeIconDecorativeShorter: SizingWidthHeightSemanticToken { Self.mockThemeSizeRawToken }
+    override var sizeIconDecorativeShort: SizingWidthHeightSemanticToken { Self.mockThemeSizeRawToken }
+    override var sizeIconDecorativeMedium: SizingWidthHeightSemanticToken { Self.mockThemeSizeRawToken }
+    override var sizeIconDecorativeTall: SizingWidthHeightSemanticToken { Self.mockThemeSizeRawToken }
+    override var sizeIconDecorativeTaller: SizingWidthHeightSemanticToken { Self.mockThemeSizeRawToken }
+    override var sizeIconDecorativeTallest: SizingWidthHeightSemanticToken { Self.mockThemeSizeRawToken }
 
     // MARK: Semantic token - Sizing - Width height - Icon component
 
-    override var sizeWidthHeightIsLabelSmallShort: SizingWidthHeightSemanticToken { Self.mockThemeSizeRawToken }
-    override var sizeWidthHeightIsLabelSmallMedium: SizingWidthHeightSemanticToken { Self.mockThemeSizeRawToken }
-    override var sizeWidthHeightIsLabelSmallTall: SizingWidthHeightSemanticToken { Self.mockThemeSizeRawToken }
-    override var sizeWidthHeightIsLabelMediumShort: SizingWidthHeightSemanticToken { Self.mockThemeSizeRawToken }
-    override var sizeWidthHeightIsLabelMediumMedium: SizingWidthHeightSemanticToken { Self.mockThemeSizeRawToken }
-    override var sizeWidthHeightIsLabelMediumTall: SizingWidthHeightSemanticToken { Self.mockThemeSizeRawToken }
-    override var sizeWidthHeightIsLabelLargeShorter: SizingWidthHeightSemanticToken { Self.mockThemeSizeRawToken }
-    override var sizeWidthHeightIsLabelLargeShort: SizingWidthHeightSemanticToken { Self.mockThemeSizeRawToken }
-    override var sizeWidthHeightIsLabelLargeMedium: SizingWidthHeightSemanticToken { Self.mockThemeSizeRawToken }
-    override var sizeWidthHeightIsLabelLargeTall: SizingWidthHeightSemanticToken { Self.mockThemeSizeRawToken }
-    override var sizeWidthHeightIsLabelLargeTaller: SizingWidthHeightSemanticToken { Self.mockThemeSizeRawToken }
-    override var sizeWidthHeightIsLabelXLargeShort: SizingWidthHeightSemanticToken { Self.mockThemeSizeRawToken }
-    override var sizeWidthHeightIsLabelXLargeMedium: SizingWidthHeightSemanticToken { Self.mockThemeSizeRawToken }
-    override var sizeWidthHeightIsLabelXLargeTall: SizingWidthHeightSemanticToken { Self.mockThemeSizeRawToken }
+    override var sizeIconWithTypeHeadingSmallShort: SizingWidthHeightSemanticToken { Self.mockThemeSizeRawToken }
+    override var sizeIconWithTypeHeadingSmallMedium: SizingWidthHeightSemanticToken { Self.mockThemeSizeRawToken }
+    override var sizeIconWithTypeHeadingSmallTall: SizingWidthHeightSemanticToken { Self.mockThemeSizeRawToken }
+    override var sizeIconWithTypeHeadingMediumShort: SizingWidthHeightSemanticToken { Self.mockThemeSizeRawToken }
+    override var sizeIconWithTypeHeadingMediumMedium: SizingWidthHeightSemanticToken { Self.mockThemeSizeRawToken }
+    override var sizeIconWithTypeHeadingTall: SizingWidthHeightSemanticToken { Self.mockThemeSizeRawToken }
+    override var sizeIconWithTypeHeadingLargeShort: SizingWidthHeightSemanticToken { Self.mockThemeSizeRawToken }
+    override var sizeIconWithTypeHeadingLargeMedium: SizingWidthHeightSemanticToken { Self.mockThemeSizeRawToken }
+    override var sizeIconWithTypeHeadingTall: SizingWidthHeightSemanticToken { Self.mockThemeSizeRawToken }
+    override var sizeIconWithTypeHeadingXLargeShort: SizingWidthHeightSemanticToken { Self.mockThemeSizeRawToken }
+    override var sizeIconWithTypeHeadingXLargeMedium: SizingWidthHeightSemanticToken { Self.mockThemeSizeRawToken }
+    override var sizeIconWithTypeHeadingXLargeTall: SizingWidthHeightSemanticToken { Self.mockThemeSizeRawToken }
+    override var sizeIconWithTypeLabelXLargeMedium: SizingWidthHeightSemanticToken { Self.mockThemeSizeRawToken }
+    override var sizeIconWithTypeLabelXLargeTall: SizingWidthHeightSemanticToken { Self.mockThemeSizeRawToken }
 
     // MARK: Semantic token - Sizing - Width height - Icon typography - Heading
 
@@ -64,26 +64,26 @@ extension MockTheme {
 
     // MARK: Semantic token - Sizing - Width height - Icon typography - Body
 
-    override var sizeWidthHeightIconIsBodySmallShort: SizingCompositeSemanticToken { Self.mockThemeSizeCompositeToken }
-    override var sizeWidthHeightIconIsBodySmallMedium: SizingCompositeSemanticToken { Self.mockThemeSizeCompositeToken }
-    override var sizeWidthHeightIconIsBodySmallTall: SizingCompositeSemanticToken { Self.mockThemeSizeCompositeToken }
-    override var sizeWidthHeightIconIsBodyMediumShort: SizingCompositeSemanticToken { Self.mockThemeSizeCompositeToken }
-    override var sizeWidthHeightIconIsBodyMediumMedium: SizingCompositeSemanticToken { Self.mockThemeSizeCompositeToken }
-    override var sizeWidthHeightIconIsBodyMediumTall: SizingCompositeSemanticToken { Self.mockThemeSizeCompositeToken }
-    override var sizeWidthHeightIconIsBodyLargeShort: SizingCompositeSemanticToken { Self.mockThemeSizeCompositeToken }
-    override var sizeWidthHeightIconIsBodyLargeMedium: SizingCompositeSemanticToken { Self.mockThemeSizeCompositeToken }
-    override var sizeWidthHeightIconIsBodyLargeTall: SizingCompositeSemanticToken { Self.mockThemeSizeCompositeToken }
+    override var sizeIconWithTypeBodySmallShort: SizingCompositeSemanticToken { Self.mockThemeSizeCompositeToken }
+    override var sizeIconWithTypeBodySmallMediuù: SizingCompositeSemanticToken { Self.mockThemeSizeCompositeToken }
+    override var sizeIconWithTypeBodySmallTall: SizingCompositeSemanticToken { Self.mockThemeSizeCompositeToken }
+    override var sizeIconWithTypeBodyMediumShort: SizingCompositeSemanticToken { Self.mockThemeSizeCompositeToken }
+    override var sizeIconWithTypeBodyMediumMedium: SizingCompositeSemanticToken { Self.mockThemeSizeCompositeToken }
+    override var sizeIconWithTypeBodyMediumTall: SizingCompositeSemanticToken { Self.mockThemeSizeCompositeToken }
+    override var sizeIconWithTypeBodyLargeShort: SizingCompositeSemanticToken { Self.mockThemeSizeCompositeToken }
+    override var sizeIconWithTypeBodyLargeMedium: SizingCompositeSemanticToken { Self.mockThemeSizeCompositeToken }
+    override var sizeIconWithTypeBodyLargeTall: SizingCompositeSemanticToken { Self.mockThemeSizeCompositeToken }
 
     // MARK: Semantic token - Sizing - Max width
 
-    override var sizeMaxWidthTypographyDisplaySmall: SizingCompositeSemanticToken { Self.mockThemeSizeCompositeToken }
-    override var sizeMaxWidthTypographyDisplayMedium: SizingCompositeSemanticToken { Self.mockThemeSizeCompositeToken }
-    override var sizeMaxWidthTypographyDisplayLarge: SizingCompositeSemanticToken { Self.mockThemeSizeCompositeToken }
-    override var sizeMaxWidthTypographyHeadingSmall: SizingCompositeSemanticToken { Self.mockThemeSizeCompositeToken }
-    override var sizeMaxWidthTypographyHeadingMedium: SizingCompositeSemanticToken { Self.mockThemeSizeCompositeToken }
-    override var sizeMaxWidthTypographyHeadingLarge: SizingCompositeSemanticToken { Self.mockThemeSizeCompositeToken }
-    override var sizeMaxWidthTypographyHeadingXLarge: SizingCompositeSemanticToken { Self.mockThemeSizeCompositeToken }
-    override var sizeMaxWidthTypographyBodySmall: SizingCompositeSemanticToken { Self.mockThemeSizeCompositeToken }
-    override var sizeMaxWidthTypographyBodyMedium: SizingCompositeSemanticToken { Self.mockThemeSizeCompositeToken }
-    override var sizeMaxWidthTypographyBodyLarge: SizingCompositeSemanticToken { Self.mockThemeSizeCompositeToken }
+    override var sizeMaxWidthTypeDisplaySmall: SizingCompositeSemanticToken { Self.mockThemeSizeCompositeToken }
+    override var sizeMaxWidthTypeDisplayMedium: SizingCompositeSemanticToken { Self.mockThemeSizeCompositeToken }
+    override var sizeMaxWidthTypeDisplayLarge: SizingCompositeSemanticToken { Self.mockThemeSizeCompositeToken }
+    override var sizeMaxWidthTypeHeadingSmall: SizingCompositeSemanticToken { Self.mockThemeSizeCompositeToken }
+    override var sizeMaxWidthTypeHeadingMedium: SizingCompositeSemanticToken { Self.mockThemeSizeCompositeToken }
+    override var sizeMaxWidthTypeHeadingLarge: SizingCompositeSemanticToken { Self.mockThemeSizeCompositeToken }
+    override var sizeMaxWidthTypeHeadingXLarge: SizingCompositeSemanticToken { Self.mockThemeSizeCompositeToken }
+    override var sizeMaxWidthTypeBodySmall: SizingCompositeSemanticToken { Self.mockThemeSizeCompositeToken }
+    override var sizeMaxWidthTypeBodyMedium: SizingCompositeSemanticToken { Self.mockThemeSizeCompositeToken }
+    override var sizeMaxWidthTypeBodyLarge: SizingCompositeSemanticToken { Self.mockThemeSizeCompositeToken }
 }