+++ OUDS/Core/Tokens/SemanticTokens/Sources/TypeAliases/SizingSemanticTokens+Aliases.swift
@@ -13,6 +13,8 @@
 
 import OUDSTokensRaw
 
+// TODO: #112 - Rename SizingWidthHeightSemanticToken to SizingSemanticToken
+
 /// Basically a size semantic token for width and height is a dimension raw token, it has the same final type, to keep grammar clean and clear with design system grammar.
 public typealias SizingWidthHeightSemanticToken = DimensionRawToken